Improved provisioning script

+++ scripts/dev/provision_vagrant.sh
@@ -1,9 +1,19 @@
 #!/bin/bash
 
+if [ -e /etc/akvo_provisioned ]
+then
+    echo "Already bootstrapped, so nothing to do"
+    exit 0
+fi
+
 su rsr
 
-ln -s /vagrant/rsr/checkout /var/akvo/rsr/git/current
-ln -s /var/akvo/rsr/local_settings.conf /var/akvo/rsr/git/current/akvo/settings/65_puppet.conf
+if [ ! -L /var/akvo/rsr/git/checkout ]
+then
+    ln -s /vagrant/rsr/checkout /var/akvo/rsr/git/current
+fi
+
+ln -sf /var/akvo/rsr/local_settings.conf /var/akvo/rsr/git/current/akvo/settings/65_puppet.conf
 
 /var/akvo/rsr/venv/bin/pip install -r /var/akvo/rsr/git/current/scripts/deployment/pip/requirements/2_rsr.txt
